+++ src/xqt-kotlinx-lsp/commonTest/xqt/kotlinx/lsp/test/lifecycle/ClientCapabilities.kt
@@ -1,0 +1,69 @@
+// Copyright (C) 2023 Reece H. Dunn. SPDX-License-Identifier: Apache-2.0
+package xqt.kotlinx.lsp.test.lifecycle
+
+import kotlinx.serialization.json.JsonNull
+import kotlinx.serialization.json.JsonPrimitive
+import xqt.kotlinx.lsp.lifecycle.ClientCapabilities
+import xqt.kotlinx.rpc.json.serialization.UnsupportedKindTypeException
+import xqt.kotlinx.rpc.json.serialization.jsonArrayOf
+import xqt.kotlinx.rpc.json.serialization.jsonObjectOf
+import xqt.kotlinx.test.DisplayName
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFails
+
+@DisplayName("The client capabilities")
+class TheClientCapabilities {
+    @Test
+    @DisplayName("supports the non-optional properties")
+    fun supports_the_non_optional_properties() {
+        val json = jsonObjectOf()
+
+        val caps = ClientCapabilities.deserialize(json)
+        assertEquals(0, caps.size)
+
+        assertEquals(json, ClientCapabilities.serializeToJson(caps))
+    }
+
+    @Test
+    @DisplayName("supports the process ID property")
+    fun supports_the_process_id_property() {
+        val json = jsonObjectOf(
+            "test" to JsonPrimitive("lorem ipsum")
+        )
+
+        val caps = ClientCapabilities.deserialize(json)
+        assertEquals(1, caps.size)
+        assertEquals(JsonPrimitive("lorem ipsum"), caps["test"])
+
+        assertEquals(json, ClientCapabilities.serializeToJson(caps))
+    }
+
+    @Test
+    @DisplayName("throws an error if the kind type is not supported")
+    fun throws_an_error_if_the_kind_type_is_not_supported() {
+        val e1 = assertFails { ClientCapabilities.deserialize(jsonArrayOf()) }
+        assertEquals(UnsupportedKindTypeException::class, e1::class)
+        assertEquals("Unsupported kind type 'array'", e1.message)
+
+        val e2 = assertFails { ClientCapabilities.deserialize(JsonNull) }
+        assertEquals(UnsupportedKindTypeException::class, e2::class)
+        assertEquals("Unsupported kind type 'null'", e2.message)
+
+        val e3 = assertFails { ClientCapabilities.deserialize(JsonPrimitive("test")) }
+        assertEquals(UnsupportedKindTypeException::class, e3::class)
+        assertEquals("Unsupported kind type 'string'", e3.message)
+
+        val e4 = assertFails { ClientCapabilities.deserialize(JsonPrimitive(true)) }
+        assertEquals(UnsupportedKindTypeException::class, e4::class)
+        assertEquals("Unsupported kind type 'boolean'", e4.message)
+
+        val e5 = assertFails { ClientCapabilities.deserialize(JsonPrimitive(1)) }
+        assertEquals(UnsupportedKindTypeException::class, e5::class)
+        assertEquals("Unsupported kind type 'integer'", e5.message)
+
+        val e6 = assertFails { ClientCapabilities.deserialize(JsonPrimitive(1.2)) }
+        assertEquals(UnsupportedKindTypeException::class, e6::class)
+        assertEquals("Unsupported kind type 'decimal'", e6.message)
+    }
+}